requirements.txt
numpy
pandas
numba
polygon-api-client
matplotlib
scipy
tqdm
